fix(github): os TRES graficos perdiam informacao — e por motivos diferentes
Mesmo defeito de familia do Codex, mas cada tela sofria de um jeito. Medido no
payload real de 30/08 (13 dias, 12 a 871 min: 73x de alcance).

1) AGORA — sete barras diarias (ui_github.cpp:226)
   Escala linear com piso de 2px: dias de 12 e 18 min desenhavam 2px contra
   44px do dia de pico. Passa a raiz quarta -> 15 e 17px.

2) PROJETOS — faixa de rateio (ui_github.cpp:242)
   ⚠️ Aqui eu previ errado. Anunciei que o piso ESTOURAVA os 452px do
   conteiner; medindo, a soma dava 449 — FALTAVA 3px, perdidos na divisao
   inteira (438+5+4+2). O piso nem estava ativo.

   E esta faixa NAO podia receber raiz nenhuma: ela e PROPORCAO, nao
   magnitude. Se helm consome 97% do tempo, a barra dele tem de ocupar 97% da
   largura — comprimir isso destruiria a unica informacao que a faixa da.

   Dai a separacao: cxRepartir() faz so o maior-resto, sem escala, e quem
   chama decide. cxAlturaColuna() e a escala. cxColunaAlturas() e a composicao
   das duas, para os graficos empilhados.

3) JOBS — empilhado de 14 dias (ui_github.cpp:273)
   Identico ao caso do Codex: piso de 2px por segmento num grafico empilhado,
   entao a altura media QUANTOS PROJETOS rodaram, nao quanto consumiram.

TIPOS: cxColunaAlturas/cxRepartir passam a uint32_t. `GhItem.min` ja era
uint32_t de proposito — 65535 minutos sao 1092 horas, alcancaveis num ciclo
longo, e a assinatura em uint16_t truncaria 70000 para 4464, fazendo a maior
fatia desenhar MENOR que a menor. Ha teste para isso.

Suite do host: 8/8, com 13 assercoes novas. Binario 2.025.318 B (12%), RAM
estatica inalterada.

⚠️ NAO CORRIGIDO, porque e modelagem e nao desenho: o `diario` do GitHub traz
13 dias, nao 14 — o servidor OMITE dias sem atividade (28/08 nao existe no
payload). O eixo horizontal, portanto, nao e uniforme no tempo: dois dias
vizinhos no grafico podem estar separados por um buraco.

+++ firmware/claude_stick/ui_refresh.cpp
@@ -383,7 +383,7 @@ void refresh_codex_analytics() {
         // do dia e maior-resto nos segmentos. O piso de 2px por segmento saiu —
         // num grafico EMPILHADO ele somava 2px por origem minuscula, achatava
         // 73% dos dados e ainda estourava o teto (62px num grafico de 58).
-        uint16_t vv[CXAN_SURF];
+        uint32_t vv[CXAN_SURF];
         int alt[CXAN_SURF];
         for (int s = 0; s < CXAN_SURF; s++)
           vv[s] = (s < g_codex.nSurfOrder) ? g_codex.day[i].v[s] : 0;
